feat: remove deprecated telemetry provider
